src/administration/serializers.py
from allauth.account.adapter import get_adapter
from allauth.account.utils import setup_user_email
from django.utils.translation import ugettext_lazy as _
from rest_framework import serializers

from administration.models import User

class UserDetailsShortSerializer(serializers.ModelSerializer):

    """
    User model w/o password
    """
    class Meta:
        model = User
        fields = ('email', 'first_name', 'last_name')
        read_only_fields = ('email', )


class JWTSerializer(serializers.Serializer):
    """
    Serializer for JWT authentication.
    """
    token = serializers.CharField()
    user = UserDetailsShortSerializer()